Move IntraDayDividendSplit to FactorFileGenerator
This is a type used internally by the generator and is not intended to
be used anywhere, so moved it to a private nested class.
