--- Common/Data/Market/IntraDayDividendSplit.cs
@@ -1,42 +1,0 @@
-﻿/*
- * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
- * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-using System;
-
-namespace QuantConnect.Data.Market
-{
-    /// <summary>
-    /// Represents a intra day divdend split.
-    /// </summary>
-    public class IntraDayDividendSplit : BaseData
-    {
-        public Split Split { get; private set; }
-
-        public Dividend Dividend { get; private set; }
-
-        public IntraDayDividendSplit(Split split, Dividend dividend)
-        {
-            if (split == null) 
-                throw new ArgumentNullException("split");
-            if (dividend == null)
-                throw new ArgumentNullException("dividend");
-
-
-            Split = split;
-            Dividend = dividend;
-            Time = Split.Time;
-        }
-    }
-}
